backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
firebase-admin==6.6.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
requests==2.31.0

